Firebase ID token verification

+++ src/main/java/com/avpc/restfulcontrollers/VehicleController.java
@@ -8,7 +8,9 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import utils.Constants;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -25,32 +27,43 @@ public class VehicleController {
 
     @RequestMapping(method = RequestMethod.POST)
     public void addVehicle(@RequestBody VehicleDTO vehicleParams,
-                           HttpServletResponse response) throws IOException {
-        try{
-            Vehicle vehicle = new Vehicle();
-            vehicle.setBrand(vehicleParams.getBrand());
-            vehicle.setCredential(vehicleParams.getCredential());
-            vehicle.setModel(vehicleParams.getModel());
-            vehicle.setRegistrationNumber(vehicleParams.getRegistration_number());
-            vehicleDAO.save(vehicle);
-        } catch (Exception e){
-            log.error(e.getMessage());
-            response.sendError(HttpStatus.CONFLICT.value());
+                           HttpServletRequest request, HttpServletResponse response) throws IOException {
+
+        if(request.getSession().getAttribute(Constants.UID) != null) {
+            try {
+                Vehicle vehicle = new Vehicle();
+                vehicle.setBrand(vehicleParams.getBrand());
+                vehicle.setCredential(vehicleParams.getCredential());
+                vehicle.setModel(vehicleParams.getModel());
+                vehicle.setRegistrationNumber(vehicleParams.getRegistration_number());
+                vehicleDAO.save(vehicle);
+            } catch (Exception e){
+                log.error(e.getMessage());
+                response.sendError(HttpStatus.CONFLICT.value());
+            }
+        }
+        else {
+            response.sendError(HttpStatus.UNAUTHORIZED.value());
         }
     }
 
     @RequestMapping(value ="/{vehicleId}", method = RequestMethod.GET)
     @ResponseBody
     public Vehicle findVehicle(@PathVariable(value="vehicleId") Long vehicleId,
-                                     HttpServletResponse response) throws IOException {
+                               HttpServletRequest request, HttpServletResponse response) throws IOException {
 
         Vehicle vehicle = null;
 
-        try{
-            vehicle =vehicleDAO.findOne(vehicleId);
-        } catch (IllegalArgumentException e){
-            log.error(e.getMessage());
-            response.sendError(HttpStatus.CONFLICT.value());
+        if(request.getSession().getAttribute(Constants.UID) != null) {
+            try {
+                vehicle =vehicleDAO.findOne(vehicleId);
+            } catch (IllegalArgumentException e){
+                log.error(e.getMessage());
+                response.sendError(HttpStatus.CONFLICT.value());
+            }
+        }
+        else {
+            response.sendError(HttpStatus.UNAUTHORIZED.value());
         }
 
         return vehicle;
@@ -58,16 +71,21 @@ public class VehicleController {
 
     @RequestMapping(method = RequestMethod.GET)
     @ResponseBody
-    public List<Vehicle> findVehicle(HttpServletResponse response) throws IOException {
+    public List<Vehicle> findVehicle(HttpServletRequest request, HttpServletResponse response) throws IOException {
         List<Vehicle> listVehicles = new ArrayList<>();
 
-        try{
-            vehicleDAO.findAll().forEach(member -> listVehicles.add(member));
-            return listVehicles;
+        if(request.getSession().getAttribute(Constants.UID) != null) {
+            try {
+                vehicleDAO.findAll().forEach(member -> listVehicles.add(member));
+                return listVehicles;
 
-        } catch (Exception e){
-            log.error(e.getMessage());
-            response.sendError(HttpStatus.CONFLICT.value());
+            } catch (Exception e){
+                log.error(e.getMessage());
+                response.sendError(HttpStatus.CONFLICT.value());
+            }
+        }
+        else {
+            response.sendError(HttpStatus.UNAUTHORIZED.value());
         }
 
         return listVehicles;
@@ -77,34 +95,45 @@ public class VehicleController {
     @ResponseBody
     public Vehicle updateVehicle(@PathVariable(value="vehicleId") Long vehicleId,
                                  @RequestBody VehicleDTO vehicleParams,
-                                 HttpServletResponse response) throws IOException {
+                                 HttpServletRequest request, HttpServletResponse response) throws IOException {
 
         Vehicle vehicle =  null;
 
-        try{
-            vehicle = vehicleDAO.findOne(vehicleId);
-            vehicle.setBrand(vehicleParams.getBrand());
-            vehicle.setCredential(vehicleParams.getCredential());
-            vehicle.setModel(vehicleParams.getModel());
-            vehicle.setRegistrationNumber(vehicleParams.getRegistration_number());
-            vehicleDAO.save(vehicle);
+        if(request.getSession().getAttribute(Constants.UID) != null) {
+            try {
+                vehicle = vehicleDAO.findOne(vehicleId);
+                vehicle.setBrand(vehicleParams.getBrand());
+                vehicle.setCredential(vehicleParams.getCredential());
+                vehicle.setModel(vehicleParams.getModel());
+                vehicle.setRegistrationNumber(vehicleParams.getRegistration_number());
+                vehicleDAO.save(vehicle);
 
-        } catch (Exception e){
-            log.error(e.getMessage());
-            response.sendError(HttpStatus.CONFLICT.value());
+            } catch (Exception e){
+                log.error(e.getMessage());
+                response.sendError(HttpStatus.CONFLICT.value());
+            }
+        }
+        else {
+            response.sendError(HttpStatus.UNAUTHORIZED.value());
         }
 
         return vehicle;
     }
 
     @RequestMapping(value ="/{vehicleId}", method = RequestMethod.DELETE)
-    public void deleteMember(@PathVariable(value="vehicleId") Long vehicleId,
-                               HttpServletResponse response) throws IOException{
-        try{
-            vehicleDAO.delete(vehicleId);
-        } catch (Exception e){
-            log.error(e.getMessage());
-            response.sendError(HttpStatus.CONFLICT.value());
+    public void deleteVehicle(@PathVariable(value="vehicleId") Long vehicleId,
+                             HttpServletRequest request, HttpServletResponse response) throws IOException{
+
+        if(request.getSession().getAttribute(Constants.UID) != null) {
+            try {
+                vehicleDAO.delete(vehicleId);
+            } catch (Exception e){
+                log.error(e.getMessage());
+                response.sendError(HttpStatus.CONFLICT.value());
+            }
+        }
+        else {
+            response.sendError(HttpStatus.UNAUTHORIZED.value());
         }
     }
 }
